implementação do delete
