+++ src/app/components/post_card/styles.module.scss
@@ -166,3 +166,4 @@
 }
 
 
+
